move Column classes to LogAnalyzer\Database\Column
