+++ src/CollectionBuilder.php
@@ -7,7 +7,7 @@ use LogAnalyzer\CollectionBuilder\LogFiles\LogFile;
 use LogAnalyzer\CollectionBuilder\Parser\ApacheLogParser;
 use LogAnalyzer\CollectionBuilder\Parser\LtsvParser;
 use LogAnalyzer\CollectionBuilder\Parser\ParserInterface;
-use LogAnalyzer\Database\ColumnFactory;
+use LogAnalyzer\Database\Column\ColumnFactory;
 use LogAnalyzer\Database\DatabaseInterface;
 use LogAnalyzer\Database\ColumnarDatabase;
