+++ src/Database/Column/FileStorageColumn.php
@@ -1,5 +1,7 @@
 <?php
-namespace LogAnalyzer\Database;
+namespace LogAnalyzer\Database\Column;
+
+use LogAnalyzer\Database\Column\ColumnInterface;
 
 class FileStorageColumn implements ColumnInterface
 {
