+++ src/Database/Column/InMemoryColumn.php
@@ -1,5 +1,7 @@
 <?php
-namespace LogAnalyzer\Database;
+namespace LogAnalyzer\Database\Column;
+
+use LogAnalyzer\Database\Column\ColumnInterface;
 
 class InMemoryColumn implements ColumnInterface
 {
